API/requirements.txt
amqp==2.6.1
arabic-reshaper==2.1.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.21.45
botocore==1.24.45
celery==4.4.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.1
django-environ==0.8.1
django-mathfilters==1.0.0
django-storages==1.12.3
djangorestframework==3.13.1
future==0.18.2
html5lib==1.1
idna==3.3
jmespath==1.0.0
kombu==4.6.11
packaging==21.3
pdfkit==1.0.0
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.4
reportlab==3.6.7
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
twilio==7.6.0
urllib3==1.26.8
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.0.0
wrapt==1.13.3
xhtml2pdf==0.2.5
